refactor: код по закачки текущего фильма перещен в хук use-loaded-movie
