+++ src/components/pages/player/player.jsx
@@ -1,26 +1,16 @@
-import React, {useEffect, useState} from "react";
+import React, {useState} from "react";
 import MainLayout from "../../layouts/main-layout/main-layout";
 import VideoPlayer from "../../sections/video-player/video-player";
-import {useDispatch, useSelector} from "react-redux";
 import MoviesList from "../../blocks/movies-list/movies-list";
 import Loader from "../../blocks/loader/loader";
 import {useHistory} from "react-router-dom";
-import {fetchCurrentMovie} from "../../../store/movie-data/movie-data-api-actions";
+import useLoadedMovie from "../../../hooks/use-loaded-movie";
 
 const Player = ({id}) => {
   const history = useHistory();
-  const isCurrentMovieLoaded = useSelector((state) => state.isCurrentMovieLoaded);
-  const currentMovie = useSelector((state) => state.currentMovie);
-  const dispatch = useDispatch();
-
-  const {videoLink} = currentMovie;
   const [isPlaying, setIsPlaying] = useState(true);
-
-  useEffect(() => {
-    if (!isCurrentMovieLoaded || !currentMovie) {
-      dispatch(fetchCurrentMovie(id));
-    }
-  }, [id, isCurrentMovieLoaded, currentMovie]);
+  const [isCurrentMovieLoaded, currentMovie] = useLoadedMovie(id);
+  const {videoLink} = currentMovie;
 
   return isCurrentMovieLoaded ?
     (<MainLayout>
